// src/api/model/votacion.ts
/**
 * Plenos API
 * API para gestionar consultas de usuarios.
 *
 * OpenAPI spec version: 3.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface Votacion { 
    id: number;
    fecha: Date;
    infoExtra: string;
    favor: number;
    contra: number;
    blanco: number;
    idPunto: number;
}